New design with benchmarks and proper testing vs the comparable algorithm now.
This version of the new design is basically ready to be implemented for 
the full package now. A major issue is that the interface for 
NetworkDynamics has changed. We now require the parameters in order to 
calculate the JacVecOperator. Maybe we can keep the old interface around 
(dispatch on the number of parameters)...